1.6.0 dev (#149)
* feature: support nsexec


* chore: update version to 1.6.0


* fix: process hang exp


* feature: improvement compile, remove parse java models


* feature: update cri version
