+++ Makefile
@@ -7,7 +7,7 @@ GO=env $(GO_ENV) $(GO_MODULE) go
 UNAME := $(shell uname)
 
 ifeq ($(BLADE_VERSION), )
-	BLADE_VERSION=1.5.0
+	BLADE_VERSION=1.6.0
 endif
 ifeq ($(BLADE_VENDOR), )
 	BLADE_VENDOR=community
@@ -27,27 +27,20 @@ VERSION_PKG=github.com/chaosblade-io/chaosblade-operator/version
 GO_X_FLAGS=-X=$(VERSION_PKG).CombinedVersion=$(BLADE_VERSION),$(BLADE_VENDOR)
 GO_FLAGS=-ldflags $(GO_X_FLAGS)
 
-# cache downloaded file
-CACHE_PATH=build/cache
-DOWNLOAD_URL=https://chaosblade.oss-cn-hangzhou.aliyuncs.com/agent/github/${BLADE_VERSION}
-CHAOSBLADE_FILE=chaosblade-${BLADE_VERSION}-linux-amd64.tar.gz
-CHAOSBLADE_UNZIP_DIR=$(CACHE_PATH)/chaosblade-${BLADE_VERSION}
-CHAOSBLADE_PATH=$(CACHE_PATH)/chaosblade
-
 ifeq ($(GOOS), linux)
 	GO_FLAGS=-ldflags="-linkmode external -extldflags -static $(GO_X_FLAGS)"
 endif
 
 build: build_yaml build_fuse
 
-build_all: pre_build pre_chaosblade build docker-build
+build_all: pre_build build docker-build
 
 docker-build:
 	GOOS="linux" GOARCH="amd64" go build $(GO_FLAGS) -o build/_output/bin/chaosblade-operator cmd/manager/main.go
 	docker build -f build/Dockerfile -t chaosblade-operator:${BLADE_VERSION} .
 
 #operator-sdk 0.19.0 build
-build_all_operator: pre_build pre_chaosblade build build_image
+build_all_operator: pre_build build build_image
 build_image:
 	operator-sdk build --go-build-args="$(GO_FLAGS)" chaosblade-operator:${BLADE_VERSION}
 
@@ -60,26 +53,16 @@ build_linux:
 		-w /go/src/github.com/chaosblade-io/chaosblade-operator \
 		chaosblade-operator-build-musl:latest
 
-pre_chaosblade:
-ifneq ($(CHAOSBLADE_PATH), $(wildcard $(CHAOSBLADE_PATH)))
-	wget "$(DOWNLOAD_URL)/$(CHAOSBLADE_FILE)" -O $(CACHE_PATH)/$(CHAOSBLADE_FILE)
-	tar zxvf $(CACHE_PATH)/$(CHAOSBLADE_FILE) -C $(CACHE_PATH)
-	mv $(CHAOSBLADE_UNZIP_DIR) $(CHAOSBLADE_PATH)
-	rm -rf $(CACHE_PATH)/$(CHAOSBLADE_FILE)
-endif
-
 pre_build:
-	mkdir -p $(BUILD_TARGET_BIN) $(BUILD_TARGET_YAML) $(CACHE_PATH)
+	mkdir -p $(BUILD_TARGET_BIN) $(BUILD_TARGET_YAML)
 
 build_spec_yaml: build/spec.go
-	$(GO) run $< $(OS_YAML_FILE_PATH) $(CHAOSBLADE_PATH)/yaml/chaosblade-jvm-spec-$(BLADE_VERSION).yaml
-	cp $(OS_YAML_FILE_PATH) $(CHAOSBLADE_PATH)/yaml/
+	$(GO) run $< $(OS_YAML_FILE_PATH)
 
-build_yaml: pre_build pre_chaosblade build_spec_yaml
+build_yaml: pre_build build_spec_yaml
 
 build_fuse:
 	$(GO) build $(GO_FLAGS) -o $(BUILD_TARGET_BIN)/chaos_fuse cmd/hookfs/main.go
-	cp $(BUILD_TARGET_BIN)/chaos_fuse $(CHAOSBLADE_PATH)/bin/
 
 # test
 test:
@@ -87,5 +70,5 @@ test:
 # clean all build result
 clean:
 	go clean ./...
-	rm -rf $(BUILD_TARGET) $(CACHE_PATH)
+	rm -rf $(BUILD_TARGET)
 	rm -rf $(BUILD_IMAGE_PATH)/$(BUILD_TARGET_DIR_NAME)
