+++ deploy/helm/chaosblade-operator/Chart.yaml
@@ -1,6 +1,6 @@
 apiVersion: v1
-appVersion: "1.5.0"
+appVersion: "1.6.0"
 description: ChaosBlade Operator
 name: chaosblade-operator
-version: 1.5.0
+version: 1.6.0
 home: https://github.com/chaosblade-io
