+++ deploy/helm/chaosblade-operator/values.yaml
@@ -3,7 +3,7 @@
 # chaosblade-operator
 operator:
   repository: chaosbladeio/chaosblade-operator
-  version: 1.5.0
+  version: 1.6.0
   # image.pullPolicy: must be Always|IfNotPresent|Never
   pullPolicy: IfNotPresent
   # qps of kubernetes client
@@ -12,7 +12,7 @@ operator:
 
 blade:
   repository: chaosbladeio/chaosblade-tool
-  version: 1.5.0
+  version: 1.6.0
   pullPolicy: IfNotPresent
   downloadUrl: ""
