+++ exec/model/executor.go
@@ -24,174 +24,25 @@ import (
 	"path"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
 
+	"github.com/chaosblade-io/chaosblade-exec-cri/exec"
 	"github.com/chaosblade-io/chaosblade-exec-cri/exec/container"
-	"github.com/chaosblade-io/chaosblade-exec-docker/exec"
 	"github.com/chaosblade-io/chaosblade-spec-go/util"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/chaosblade-io/chaosblade-spec-go/spec"
 	"github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	pkglabels "k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	cli "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/chaosblade-io/chaosblade-operator/channel"
 	"github.com/chaosblade-io/chaosblade-operator/pkg/apis/chaosblade/v1alpha1"
 	"github.com/chaosblade-io/chaosblade-operator/pkg/runtime/chaosblade"
-	"github.com/chaosblade-io/chaosblade-operator/version"
 )
 
-type ExperimentIdentifierInPod struct {
-	ContainerObjectMeta
-	Command string
-	Error   string
-	Code    int32
-	// For daemonset
-	ChaosBladePodName       string
-	ChaosBladeNamespace     string
-	ChaosBladeContainerName string
-}
-
-type ExecCommandInPodExecutor struct {
-	Client *channel.Client
-}
-
-func (e *ExecCommandInPodExecutor) Name() string {
-	return "execInPod"
-}
-
-func (e *ExecCommandInPodExecutor) SetChannel(channel spec.Channel) {
-}
-
-func (e *ExecCommandInPodExecutor) Exec(uid string, ctx context.Context, expModel *spec.ExpModel) *spec.Response {
-	return e.execInMatchedPod(uid, ctx, expModel)
-}
-
-// getExperimentIdentifiers
-func (e *ExecCommandInPodExecutor) getExperimentIdentifiers(ctx context.Context, expModel *spec.ExpModel) ([]ExperimentIdentifierInPod, error) {
-	delete(expModel.ActionFlags, "uid")
-	containerObjectMetaList, err := GetContainerObjectMetaListFromContext(ctx)
-	if err != nil {
-		return []ExperimentIdentifierInPod{}, err
-	}
-	excludeFlagsFunc := ExcludeKeyFunc()
-	matchers := spec.ConvertExpMatchersToString(expModel, excludeFlagsFunc)
-	experimentId := GetExperimentIdFromContext(ctx)
-	_, destroy := spec.IsDestroy(ctx)
-
-	isDockerNetwork := expModel.ActionFlags[IsDockerNetworkFlag.Name] == "true"
-	UseSidecarContainerNetwork := expModel.ActionFlags[UseSidecarContainerNetworkFlag.Name] == "true"
-	isContainerSelfTarget := expModel.Target == "container"
-	isContainerNetworkTarget := expModel.Target == "network"
-	isNodeScope := expModel.Scope == "node"
-	if isNodeScope {
-		return e.getNodeExperimentIdentifiers(experimentId, expModel, containerObjectMetaList, matchers, destroy)
-	}
-	if isContainerSelfTarget || (isContainerNetworkTarget && (isDockerNetwork || UseSidecarContainerNetwork)) {
-		if version.CheckVerisonHaveCriCommand() || containerObjectMetaList[0].ContainerRuntime == container.ContainerdRuntime {
-			return e.getCriExperimentIdentifiers(experimentId, expModel, containerObjectMetaList, matchers, destroy, isContainerNetworkTarget)
-		}
-		return e.getDockerExperimentIdentifiers(experimentId, expModel, containerObjectMetaList, matchers, destroy, isContainerNetworkTarget)
-	}
-	if destroy {
-		return e.generateDestroyCommands(experimentId, expModel, containerObjectMetaList, matchers)
-	}
-	return e.generateCreateCommands(experimentId, expModel, containerObjectMetaList, matchers)
-}
-
-// execInMatchedPod will execute the experiment in the target pod
-func (e *ExecCommandInPodExecutor) execInMatchedPod(uid string, ctx context.Context, expModel *spec.ExpModel) *spec.Response {
-	logrusField := logrus.WithField("experiment", GetExperimentIdFromContext(ctx))
-	experimentStatus := v1alpha1.ExperimentStatus{
-		ResStatuses: make([]v1alpha1.ResourceStatus, 0),
-	}
-	experimentIdentifiers, err := e.getExperimentIdentifiers(ctx, expModel)
-	if err != nil {
-		logrusField.Errorf("get experiment identifiers failed, err: %s", err.Error())
-		return spec.ResponseFailWithResult(spec.GetIdentifierFailed,
-			v1alpha1.CreateFailExperimentStatus(err.Error(), []v1alpha1.ResourceStatus{}),
-			err)
-	}
-	logrusField.Infof("experiment identifiers: %v", experimentIdentifiers)
-
-	statuses := experimentStatus.ResStatuses
-	success := true
-	_, isDestroy := spec.IsDestroy(ctx)
-	updateResultLock := &sync.Mutex{}
-
-	execCommandInPod := func(i int) {
-		execSuccess := true
-		identifier := experimentIdentifiers[i]
-
-		rsStatus := v1alpha1.ResourceStatus{
-			Kind:       expModel.Scope,
-			Identifier: identifier.GetIdentifier(),
-			Id:         identifier.Id,
-		}
-
-		if identifier.Error != "" {
-			rsStatus.CreateFailResourceStatus(identifier.Error, spec.K8sExecFailed.Code)
-			execSuccess = false
-		} else if identifier.PodName != "" {
-			// check if pod exist
-			pod := &v1.Pod{}
-			err := e.Client.Get(context.TODO(), types.NamespacedName{Namespace: identifier.Namespace,
-				Name: identifier.PodName}, pod)
-			if err != nil {
-				if apierrors.IsNotFound(err) {
-					// pod if not exist, the execution is considered successful.
-					msg := fmt.Sprintf("pod: %s in %s not found, skip to execute command in it",
-						identifier.PodName, identifier.Namespace)
-					logrusField.Warningln(msg)
-					rsStatus.CreateSuccessResourceStatus()
-					rsStatus.Error = msg
-					success = true
-				} else {
-					// if get pod error, the execution is considered failure.
-					msg := fmt.Sprintf("get pod: %s in %s error",
-						identifier.PodName, identifier.Namespace)
-					rsStatus.CreateFailResourceStatus(msg, spec.K8sExecFailed.Code)
-					execSuccess = false
-				}
-			}
-		}
-		if execSuccess {
-			logrusField.Infof("execute identifier: %+v", identifier)
-			execSuccess, rsStatus = e.execCommands(isDestroy, rsStatus, identifier)
-		}
-		updateResultLock.Lock()
-		statuses = append(statuses, rsStatus)
-		// If false occurs once, the result is false.
-		success = success && execSuccess
-		updateResultLock.Unlock()
-	}
-
-	ParallelizeExec(len(experimentIdentifiers), execCommandInPod)
-
-	logrusField.Infof("success: %t, statuses: %+v", success, statuses)
-	if success {
-		experimentStatus.State = v1alpha1.SuccessState
-	} else {
-		experimentStatus.State = v1alpha1.ErrorState
-		if len(statuses) == 0 {
-			experimentStatus.Error = "the resources not found"
-		} else {
-			experimentStatus.Error = "see resStatus for the error details"
-		}
-	}
-	experimentStatus.Success = success
-	experimentStatus.ResStatuses = append(experimentStatus.ResStatuses, statuses...)
-
-	checkExperimentStatus(ctx, expModel, statuses, experimentIdentifiers, e)
-	return spec.ReturnResultIgnoreCode(experimentStatus)
-}
-
-func checkExperimentStatus(ctx context.Context, expModel *spec.ExpModel, statuses []v1alpha1.ResourceStatus, identifiers []ExperimentIdentifierInPod, e *ExecCommandInPodExecutor) {
+func checkExperimentStatus(ctx context.Context, expModel *spec.ExpModel, statuses []v1alpha1.ResourceStatus, identifiers []ExperimentIdentifierInPod, client *channel.Client) {
 	tt := expModel.ActionFlags["timeout"]
 	if _, ok := spec.IsDestroy(ctx); !ok && tt != "" && len(statuses) > 0 {
 		experimentId := GetExperimentIdFromContext(ctx)
@@ -230,7 +81,7 @@ func checkExperimentStatus(ctx context.Context, expModel *spec.ExpModel, statuse
 							podNamespace = identifier.ChaosBladeNamespace
 							containerName = identifier.ChaosBladeContainerName
 						}
-						response := e.Client.Exec(&channel.ExecOptions{
+						response := client.Exec(&channel.ExecOptions{
 							StreamOptions: channel.StreamOptions{
 								ErrDecoder: func(bytes []byte) interface{} {
 									content := string(bytes)
@@ -263,11 +114,10 @@ func checkExperimentStatus(ctx context.Context, expModel *spec.ExpModel, statuse
 
 					if isDestroyed {
 						logrus.Info("The experiment was destroyed, ExperimentId: ", experimentId)
-						cli := e.Client.Client
 						objectMeta := metav1.ObjectMeta{Name: experimentId}
-						err := cli.Delete(context.TODO(), &v1alpha1.ChaosBlade{
+						err :=  client.Client.Delete(context.TODO(), &v1alpha1.ChaosBlade{
 							TypeMeta: metav1.TypeMeta{
-								APIVersion: "chaosblade.io/v1alpha1",
+								APIVersion: "chaosbladio/v1alpha1",
 								Kind:       "ChaosBlade",
 							},
 							ObjectMeta: objectMeta,
@@ -284,8 +134,8 @@ func checkExperimentStatus(ctx context.Context, expModel *spec.ExpModel, statuse
 	}
 }
 
-func (e *ExecCommandInPodExecutor) execCommands(isDestroy bool, rsStatus v1alpha1.ResourceStatus,
-	identifier ExperimentIdentifierInPod) (bool, v1alpha1.ResourceStatus) {
+func execCommands(isDestroy bool, rsStatus v1alpha1.ResourceStatus,
+	identifier ExperimentIdentifierInPod, client *channel.Client) (bool, v1alpha1.ResourceStatus) {
 	success := false
 	// handle chaos experiments using daemonset mode
 	podName := identifier.PodName
@@ -296,7 +146,7 @@ func (e *ExecCommandInPodExecutor) execCommands(isDestroy bool, rsStatus v1alpha
 		podNamespace = identifier.ChaosBladeNamespace
 		containerName = identifier.ChaosBladeContainerName
 	}
-	response := e.Client.Exec(&channel.ExecOptions{
+	response := client.Exec(&channel.ExecOptions{
 		StreamOptions: channel.StreamOptions{
 			IOStreams: channel.IOStreams{
 				Out:    bytes.NewBuffer([]byte{}),
@@ -332,8 +182,8 @@ func (e *ExecCommandInPodExecutor) execCommands(isDestroy bool, rsStatus v1alpha
 	return success, rsStatus
 }
 
-func (e *ExecCommandInPodExecutor) generateDestroyCommands(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string) ([]ExperimentIdentifierInPod, error) {
+func generateDestroyCommands(experimentId string, expModel *spec.ExpModel,
+	containerObjectMetaList ContainerMatchedList, matchers string, client *channel.Client) ([]ExperimentIdentifierInPod, error) {
 	command := fmt.Sprintf("%s destroy %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
 	identifiers := make([]ExperimentIdentifierInPod, 0)
 	for idx, obj := range containerObjectMetaList {
@@ -345,7 +195,7 @@ func (e *ExecCommandInPodExecutor) generateDestroyCommands(experimentId string, 
 			ContainerObjectMeta: containerObjectMetaList[idx],
 			Command:             generatedCommand,
 		}
-		resp := e.deployChaosBlade(experimentId, expModel, obj, false)
+		resp := deployChaosBlade(experimentId, expModel, obj, false, client)
 		if !resp.Success {
 			identifierInPod.Error = resp.Err
 			identifierInPod.Code = resp.Code
@@ -355,8 +205,8 @@ func (e *ExecCommandInPodExecutor) generateDestroyCommands(experimentId string, 
 	return identifiers, nil
 }
 
-func (e *ExecCommandInPodExecutor) generateCreateCommands(experimentId string, expModel *spec.ExpModel, containerObjectMetaList ContainerMatchedList,
-	matchers string) ([]ExperimentIdentifierInPod, error) {
+func generateCreateCommands(experimentId string, expModel *spec.ExpModel, containerObjectMetaList ContainerMatchedList,
+	matchers string, client *channel.Client) ([]ExperimentIdentifierInPod, error) {
 	command := fmt.Sprintf("%s create %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
 	identifiers := make([]ExperimentIdentifierInPod, 0)
 	chaosBladeOverride := expModel.ActionFlags[exec.ChaosBladeOverrideFlag.Name] == "true"
@@ -365,7 +215,7 @@ func (e *ExecCommandInPodExecutor) generateCreateCommands(experimentId string, e
 			ContainerObjectMeta: containerObjectMetaList[idx],
 			Command:             command,
 		}
-		resp := e.deployChaosBlade(experimentId, expModel, obj, chaosBladeOverride)
+		resp := deployChaosBlade(experimentId, expModel, obj, chaosBladeOverride, client)
 		if !resp.Success {
 			identifierInPod.Error = resp.Err
 			identifierInPod.Code = resp.Code
@@ -375,147 +225,18 @@ func (e *ExecCommandInPodExecutor) generateCreateCommands(experimentId string, e
 	return identifiers, nil
 }
 
-func (e *ExecCommandInPodExecutor) deployChaosBlade(experimentId string, expModel *spec.ExpModel,
-	obj ContainerObjectMeta, override bool) *spec.Response {
-	logrusField := logrus.WithField("experiment", experimentId)
-	chaosBladePath := getTargetChaosBladePath(expModel)
-	options := DeployOptions{
-		Container: obj.ContainerName,
-		Namespace: obj.Namespace,
-		PodName:   obj.PodName,
-		client:    e.Client,
-	}
-	deploy, err := getDeployMode(options, expModel)
-	if err != nil {
-		util.Errorf(experimentId, util.GetRunFuncName(), spec.ParameterLess.Sprintf(ChaosBladeDownloadUrlFlag.Name))
-		return spec.ResponseFailWithFlags(spec.ParameterLess, ChaosBladeDownloadUrlFlag.Name)
-	}
-	logrusField.Infof("deploy chaosblade under override with %t value", override)
-	chaosBladeBinPath := path.Join(chaosBladePath, "bin")
-	if err := options.CheckFileExists(chaosBladeBinPath); err != nil {
-		// create chaosblade path
-		if err := options.CreateDir(chaosBladeBinPath); err != nil {
-			util.Errorf(experimentId, util.GetRunFuncName(), fmt.Sprintf("create chaosblade dir: %s, failed! err: %s", chaosBladeBinPath, err.Error()))
-			return spec.ResponseFailWithFlags(spec.ParameterInvalidBladePathError, ChaosBladePathFlag.Name, chaosBladeBinPath, err)
-		}
-	}
-	bladePath := path.Join(chaosBladePath, "blade")
-	if override || options.CheckFileExists(bladePath) != nil {
-		if err := deploy.DeployToPod(experimentId, chaosblade.OperatorChaosBladeBlade, bladePath); err != nil {
-			util.Errorf(experimentId, util.GetRunFuncName(), fmt.Sprintf("deploy blade failed! dir: %s, err: %s", bladePath, err.Error()))
-			return spec.ResponseFailWithFlags(spec.DeployChaosBladeFailed, bladePath, err)
-		}
-	}
-	yamlPath := path.Join(chaosBladePath, "yaml")
-	if override || options.CheckFileExists(yamlPath) != nil {
-		if err := deploy.DeployToPod(experimentId, chaosblade.OperatorChaosBladeYaml, yamlPath); err != nil {
-			util.Errorf(experimentId, util.GetRunFuncName(), fmt.Sprintf("deploy yaml failed! dir: %s, err: %s", yamlPath, err.Error()))
-			return spec.ResponseFailWithFlags(spec.DeployChaosBladeFailed, yamlPath, err)
-		}
-	}
-	chaosOSPath := path.Join(chaosBladePath, "bin", "chaos_os")
-	if override || options.CheckFileExists(chaosOSPath) != nil {
-		if err := deploy.DeployToPod(experimentId, path.Join(chaosblade.OperatorChaosBladeBin, "chaos_os"), chaosOSPath); err != nil {
-			util.Errorf(experimentId, util.GetRunFuncName(), fmt.Sprintf("deploy chaos_os failed! dir: %s, err: %s", chaosOSPath, err.Error()))
-			return spec.ResponseFailWithFlags(spec.DeployChaosBladeFailed, chaosOSPath, err)
-		}
-	}
-	// copy files as needed
-	for _, program := range expModel.ActionPrograms {
-		var programFile, operatorProgramFile string
-		switch program {
-		case "java":
-			programFile = path.Join(chaosBladePath, "lib")
-			operatorProgramFile = chaosblade.OperatorChaosBladeLib
-		default:
-			programFile = path.Join(chaosBladePath, "bin", program)
-			operatorProgramFile = path.Join(chaosblade.OperatorChaosBladeBin, program)
-		}
-		if !override && options.CheckFileExists(programFile) == nil {
-			logrusField.WithField("program", programFile).Infof("program exists")
-			continue
-		}
-		err := deploy.DeployToPod(experimentId, operatorProgramFile, programFile)
-		logrusField = logrusField.WithFields(logrus.Fields{
-			"container": obj.ContainerName,
-			"pod":       obj.PodName,
-			"namespace": obj.Namespace,
-		})
-		if err != nil {
-			util.Errorf(experimentId, util.GetRunFuncName(), fmt.Sprintf("copy chaosblade to pod failed! dir: %s, err: %s", yamlPath, err.Error()))
-			return spec.ResponseFailWithFlags(spec.K8sExecFailed, "copyToPod", err)
-		}
-		logrusField.Infof("deploy %s success", programFile)
-	}
-	return spec.Success()
-}
-
-func (e *ExecCommandInPodExecutor) getNewContainerIdByPod(podName, podNamespace, containerName, experimentId string) (string, error) {
-	pod := v1.Pod{}
-	err := e.Client.Get(context.TODO(), types.NamespacedName{Namespace: podNamespace, Name: podName}, &pod)
-	if err != nil {
-		logrus.WithFields(
-			logrus.Fields{
-				"experiment":    experimentId,
-				"containerName": containerName,
-			}).Warningf("can not find the pod by %s name in %s namespace, %v", podName, podNamespace, err)
-		return "", err
-	}
-	containerStatuses := pod.Status.ContainerStatuses
-	if containerStatuses == nil {
-		return "", fmt.Errorf("cannot find containers in %s pod", podName)
-	}
-	for _, containerStatus := range containerStatuses {
-		if containerName == containerStatus.Name {
-			containerRuntime, containerLongId := TruncateContainerObjectMetaUid(containerStatus.ContainerID)
-			if len(containerLongId) > 12 {
-				if containerRuntime == container.ContainerdRuntime {
-					return containerLongId, nil
-				}
-				return containerLongId[:12], nil
-			}
-			return "", fmt.Errorf("the container %s id is illegal", containerLongId)
-		}
-	}
-	return "", fmt.Errorf("cannot find the %s container in %s pod", containerName, podName)
-}
-
-func (e *ExecCommandInPodExecutor) getDockerExperimentIdentifiers(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string, destroy, isNetworkTarget bool) ([]ExperimentIdentifierInPod, error) {
-	if isNetworkTarget {
-		matchers = fmt.Sprintf("%s --image-repo %s --image-version %s",
-			matchers, chaosblade.Constant.ImageRepoFunc(), chaosblade.Version)
-	}
-	if destroy {
-		return e.generateDestroyDockerCommands(experimentId, expModel, containerObjectMetaList, matchers, isNetworkTarget)
-	}
-	return e.generateCreateDockerCommands(experimentId, expModel, containerObjectMetaList, matchers)
-}
-
-func (e *ExecCommandInPodExecutor) getCriExperimentIdentifiers(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string, destroy, isNetworkTarget bool) ([]ExperimentIdentifierInPod, error) {
-	if isNetworkTarget {
-		matchers = fmt.Sprintf("%s --image-repo %s --image-version %s",
-			matchers, chaosblade.Constant.ImageRepoFunc(), chaosblade.Version)
-	}
-	if destroy {
-		return e.generateDestroyCriCommands(experimentId, expModel, containerObjectMetaList, matchers, isNetworkTarget)
-	}
-	return e.generateCreateCriCommands(experimentId, expModel, containerObjectMetaList, matchers)
-}
-
 // GetChaosBladeDaemonsetPodName
-func (e *ExecCommandInPodExecutor) GetChaosBladeDaemonsetPodName(nodeName string) (string, error) {
+func GetChaosBladeDaemonsetPodName(nodeName string, client *channel.Client) (string, error) {
 	podName := chaosblade.DaemonsetPodNames[nodeName]
 	if podName == "" {
-		if err := e.refreshChaosBladeDaemonsetPodNames(); err != nil {
+		if err := refreshChaosBladeDaemonsetPodNames(client); err != nil {
 			return "", err
 		}
 		return chaosblade.DaemonsetPodNames[nodeName], nil
 	}
 	// check
 	pod := v1.Pod{}
-	err := e.Client.Get(context.Background(), client.ObjectKey{
+	err := client.Get(context.Background(), cli.ObjectKey{
 		Namespace: chaosblade.DaemonsetPodNamespace,
 		Name:      podName,
 	}, &pod)
@@ -523,19 +244,19 @@ func (e *ExecCommandInPodExecutor) GetChaosBladeDaemonsetPodName(nodeName string
 		return podName, nil
 	}
 	// refresh
-	if err := e.refreshChaosBladeDaemonsetPodNames(); err != nil {
+	if err := refreshChaosBladeDaemonsetPodNames(client); err != nil {
 		return "", err
 	}
 	return chaosblade.DaemonsetPodNames[nodeName], nil
 }
 
-func (e *ExecCommandInPodExecutor) refreshChaosBladeDaemonsetPodNames() error {
+func refreshChaosBladeDaemonsetPodNames(client *channel.Client) error {
 	podList := v1.PodList{}
-	opts := client.ListOptions{
+	opts := cli.ListOptions{
 		Namespace:     chaosblade.DaemonsetPodNamespace,
 		LabelSelector: pkglabels.SelectorFromSet(chaosblade.DaemonsetPodLabels),
 	}
-	if err := e.Client.List(context.TODO(), &podList, &opts); err != nil {
+	if err := client.List(context.TODO(), &podList, &opts); err != nil {
 		return err
 	}
 	podNames := make(map[string]string, len(podList.Items))
@@ -546,148 +267,14 @@ func (e *ExecCommandInPodExecutor) refreshChaosBladeDaemonsetPodNames() error {
 	return nil
 }
 
-func (e *ExecCommandInPodExecutor) generateDestroyDockerCommands(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string, isNetworkTarget bool) ([]ExperimentIdentifierInPod, error) {
-	command := fmt.Sprintf("%s destroy docker %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
-	identifiers := make([]ExperimentIdentifierInPod, 0)
-	for idx, obj := range containerObjectMetaList {
-		daemonsetPodName, err := e.GetChaosBladeDaemonsetPodName(obj.NodeName)
-		if err != nil {
-			logrus.WithField("experiment", experimentId).
-				Errorf("get chaosblade tool pod for destroying failed on %s node, %v", obj.NodeName, err)
-			return identifiers, err
-		}
-		generatedCommand := command
-		if isNetworkTarget {
-			newContainerId, err := e.getNewContainerIdByPod(obj.PodName, obj.Namespace, obj.ContainerName, experimentId)
-			if err != nil {
-				logrus.WithField("experiment", experimentId).Errorf("generate destroy docker command failed, %v", err)
-				continue
-			}
-			generatedCommand = fmt.Sprintf("%s --container-id %s", generatedCommand, newContainerId)
-		} else {
-			if obj.Id != "" {
-				generatedCommand = fmt.Sprintf("%s --uid %s", command, obj.Id)
-			}
-			generatedCommand = fmt.Sprintf("%s --container-name %s", generatedCommand, obj.ContainerName)
-		}
-		identifierInPod := ExperimentIdentifierInPod{
-			ContainerObjectMeta:     containerObjectMetaList[idx],
-			Command:                 generatedCommand,
-			ChaosBladeContainerName: chaosblade.DaemonsetPodName,
-			ChaosBladeNamespace:     chaosblade.DaemonsetPodNamespace,
-			ChaosBladePodName:       daemonsetPodName,
-		}
-		identifiers = append(identifiers, identifierInPod)
-	}
-	return identifiers, nil
-}
-
-func (e *ExecCommandInPodExecutor) generateCreateDockerCommands(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string) ([]ExperimentIdentifierInPod, error) {
-	command := fmt.Sprintf("%s create docker %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
-
-	identifiers := make([]ExperimentIdentifierInPod, 0)
-	for idx, obj := range containerObjectMetaList {
-		daemonsetPodName, err := e.GetChaosBladeDaemonsetPodName(obj.NodeName)
-		if err != nil {
-			logrus.WithField("experiment", experimentId).
-				Errorf("get chaosblade tool pod for creating failed on %s node, %v", obj.NodeName, err)
-			return identifiers, err
-		}
-		generatedCommand := fmt.Sprintf("%s --container-id %s", command, obj.ContainerId)
-		identifierInPod := ExperimentIdentifierInPod{
-			ContainerObjectMeta:     containerObjectMetaList[idx],
-			Command:                 generatedCommand,
-			ChaosBladeContainerName: chaosblade.DaemonsetPodName,
-			ChaosBladeNamespace:     chaosblade.DaemonsetPodNamespace,
-			ChaosBladePodName:       daemonsetPodName,
-		}
-		identifiers = append(identifiers, identifierInPod)
-	}
-	return identifiers, nil
-}
-
-func (e *ExecCommandInPodExecutor) generateDestroyCriCommands(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string, isNetworkTarget bool) ([]ExperimentIdentifierInPod, error) {
-	command := fmt.Sprintf("%s destroy cri %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
-	identifiers := make([]ExperimentIdentifierInPod, 0)
-	for idx, obj := range containerObjectMetaList {
-		daemonsetPodName, err := e.GetChaosBladeDaemonsetPodName(obj.NodeName)
-		if err != nil {
-			logrus.WithField("experiment", experimentId).
-				Errorf("get chaosblade tool pod for destroying failed on %s node, %v", obj.NodeName, err)
-			return identifiers, err
-		}
-		generatedCommand := command
-		if isNetworkTarget {
-			labels := []string{
-				fmt.Sprintf("io.kubernetes.pod.name=%s", obj.PodName),
-				fmt.Sprintf("io.kubernetes.pod.namespace=%s", obj.Namespace),
-			}
-			if obj.ContainerRuntime == container.DockerRuntime {
-				labels = append(labels, "io.kubernetes.docker.type=podsandbox")
-			} else if obj.ContainerRuntime == container.ContainerdRuntime {
-				labels = append(labels, "io.cri-containerd.kind=sandbox")
-			} else {
-				logrus.WithField("experiment", experimentId).
-					Errorf("unsupported container runtime %s", obj.ContainerRuntime)
-				return identifiers, fmt.Errorf("unsupported container runtime %s", obj.ContainerRuntime)
-			}
-
-			generatedCommand = fmt.Sprintf("%s --container-label-selector %s --container-runtime %s", generatedCommand, strings.Join(labels, ","), obj.ContainerRuntime)
-		} else {
-			if obj.Id != "" {
-				generatedCommand = fmt.Sprintf("%s --uid %s", command, obj.Id)
-			}
-			generatedCommand = fmt.Sprintf("%s --container-name %s --container-runtime %s", generatedCommand, obj.ContainerName, obj.ContainerRuntime)
-		}
-		identifierInPod := ExperimentIdentifierInPod{
-			ContainerObjectMeta:     containerObjectMetaList[idx],
-			Command:                 generatedCommand,
-			ChaosBladeContainerName: chaosblade.DaemonsetPodName,
-			ChaosBladeNamespace:     chaosblade.DaemonsetPodNamespace,
-			ChaosBladePodName:       daemonsetPodName,
-		}
-		identifiers = append(identifiers, identifierInPod)
-	}
-	return identifiers, nil
-}
-
-func (e *ExecCommandInPodExecutor) generateCreateCriCommands(experimentId string, expModel *spec.ExpModel,
-	containerObjectMetaList ContainerMatchedList, matchers string) ([]ExperimentIdentifierInPod, error) {
-	command := fmt.Sprintf("%s create cri %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
-
-	identifiers := make([]ExperimentIdentifierInPod, 0)
-	for idx, obj := range containerObjectMetaList {
-		daemonsetPodName, err := e.GetChaosBladeDaemonsetPodName(obj.NodeName)
-		if err != nil {
-			logrus.WithField("experiment", experimentId).
-				Errorf("get chaosblade tool pod for creating failed on %s node, %v", obj.NodeName, err)
-			return identifiers, err
-		}
-		generatedCommand := fmt.Sprintf("%s --container-id %s --container-runtime %s", command, obj.ContainerId,
-			containerObjectMetaList[idx].ContainerRuntime)
-		identifierInPod := ExperimentIdentifierInPod{
-			ContainerObjectMeta:     containerObjectMetaList[idx],
-			Command:                 generatedCommand,
-			ChaosBladeContainerName: chaosblade.DaemonsetPodName,
-			ChaosBladeNamespace:     chaosblade.DaemonsetPodNamespace,
-			ChaosBladePodName:       daemonsetPodName,
-		}
-		identifiers = append(identifiers, identifierInPod)
-	}
-	return identifiers, nil
-}
-
-func (e *ExecCommandInPodExecutor) getNodeExperimentIdentifiers(experimentId string, expModel *spec.ExpModel, containerMatchedList ContainerMatchedList, matchers string, destroy bool) ([]ExperimentIdentifierInPod, error) {
+func getNodeExperimentIdentifiers(experimentId string, expModel *spec.ExpModel, containerMatchedList ContainerMatchedList, matchers string, destroy bool, client *channel.Client) ([]ExperimentIdentifierInPod, error) {
 	if destroy {
-		return e.generateDestroyNodeCommands(experimentId, expModel, containerMatchedList, matchers)
+		return generateDestroyNodeCommands(experimentId, expModel, containerMatchedList, matchers, client)
 	}
-	return e.generateCreateNodeCommands(experimentId, expModel, containerMatchedList, matchers)
+	return generateCreateNodeCommands(experimentId, expModel, containerMatchedList, matchers, client)
 }
 
-func (e *ExecCommandInPodExecutor) generateDestroyNodeCommands(experimentId string, expModel *spec.ExpModel, containerObjectMetaList ContainerMatchedList, matchers string) ([]ExperimentIdentifierInPod, error) {
+func generateDestroyNodeCommands(experimentId string, expModel *spec.ExpModel, containerObjectMetaList ContainerMatchedList, matchers string, client *channel.Client) ([]ExperimentIdentifierInPod, error) {
 	command := fmt.Sprintf("%s destroy %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
 	identifiers := make([]ExperimentIdentifierInPod, 0)
 	for idx, obj := range containerObjectMetaList {
@@ -695,7 +282,7 @@ func (e *ExecCommandInPodExecutor) generateDestroyNodeCommands(experimentId stri
 		if obj.Id != "" {
 			generatedCommand = fmt.Sprintf("%s --uid %s", command, obj.Id)
 		}
-		daemonsetPodName, err := e.GetChaosBladeDaemonsetPodName(obj.NodeName)
+		daemonsetPodName, err := GetChaosBladeDaemonsetPodName(obj.NodeName, client)
 		if err != nil {
 			logrus.WithField("experiment", experimentId).
 				Errorf("get chaosblade tool pod for destroying failed on %s node, %v", obj.NodeName, err)
@@ -713,11 +300,11 @@ func (e *ExecCommandInPodExecutor) generateDestroyNodeCommands(experimentId stri
 	return identifiers, nil
 }
 
-func (e *ExecCommandInPodExecutor) generateCreateNodeCommands(experimentId string, expModel *spec.ExpModel, containerObjectMetaList ContainerMatchedList, matchers string) ([]ExperimentIdentifierInPod, error) {
+func generateCreateNodeCommands(experimentId string, expModel *spec.ExpModel, containerObjectMetaList ContainerMatchedList, matchers string, client *channel.Client) ([]ExperimentIdentifierInPod, error) {
 	command := fmt.Sprintf("%s create %s %s %s", getTargetChaosBladeBin(expModel), expModel.Target, expModel.ActionName, matchers)
 	identifiers := make([]ExperimentIdentifierInPod, 0)
 	for idx, obj := range containerObjectMetaList {
-		daemonsetPodName, err := e.GetChaosBladeDaemonsetPodName(obj.NodeName)
+		daemonsetPodName, err := GetChaosBladeDaemonsetPodName(obj.NodeName, client)
 		if err != nil {
 			logrus.WithField("experiment", experimentId).
 				Errorf("get chaosblade tool pod for creating failed on %s node, %v", obj.NodeName, err)
