+++ exec/model/filter.go
@@ -25,19 +25,19 @@ import (
 	"k8s.io/api/core/v1"
 )
 
-func GetOneAvailableContainerIdFromPod(pod v1.Pod) (containerId, containerName string, err error) {
+func GetOneAvailableContainerIdFromPod(pod v1.Pod) (containerId, containerName, runtime string, err error) {
 	containerStatuses := pod.Status.ContainerStatuses
 	if containerStatuses == nil || len(containerStatuses) == 0 {
-		return "", "", fmt.Errorf("the container statues is empty in %s pod", pod.Name)
+		return "", "", "", fmt.Errorf("the container statues is empty in %s pod", pod.Name)
 	}
 	for _, containerStatus := range containerStatuses {
 		if containerStatus.State.Running == nil {
 			continue
 		}
-		_,containerId := TruncateContainerObjectMetaUid(containerStatus.ContainerID)
-		return containerId, containerStatus.Name, nil
+		runtime,containerId := TruncateContainerObjectMetaUid(containerStatus.ContainerID)
+		return containerId, containerStatus.Name, runtime,nil
 	}
-	return "", "", fmt.Errorf("cannot find a valiable container in %s pod", pod.Name)
+	return "", "", "", fmt.Errorf("cannot find a valiable container in %s pod", pod.Name)
 }
 
 func ParseLabels(labels string) map[string]string {
