+++ exec/model/osexp.go
@@ -17,7 +17,13 @@
 package model
 
 import (
-	"github.com/chaosblade-io/chaosblade-exec-os/exec"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/cpu"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/disk"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/file"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/mem"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/network"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/process"
+	"github.com/chaosblade-io/chaosblade-exec-os/exec/script"
 	"github.com/chaosblade-io/chaosblade-spec-go/spec"
 )
 
@@ -29,13 +35,13 @@ func NewOSSubResourceModelSpec() SubResourceExpModelSpec {
 	modelSpec := &OSSubResourceModelSpec{
 		BaseSubResourceExpModelSpec{
 			ExpModelSpecs: []spec.ExpModelCommandSpec{
-				exec.NewCpuCommandModelSpec(),
-				exec.NewNetworkCommandSpec(),
-				exec.NewProcessCommandModelSpec(),
-				exec.NewDiskCommandSpec(),
-				exec.NewMemCommandModelSpec(),
-				exec.NewFileCommandSpec(),
-				exec.NewScriptCommandModelSpec(),
+				cpu.NewCpuCommandModelSpec(),
+				network.NewNetworkCommandSpec(),
+				process.NewProcessCommandModelSpec(),
+				disk.NewDiskCommandSpec(),
+				mem.NewMemCommandModelSpec(),
+				file.NewFileCommandSpec(),
+				script.NewScriptCommandModelSpec(),
 			},
 		},
 	}
